requirements.txt
cycler==0.11.0
dataclasses==0.8
imageio==2.15.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-python==4.5.4.60
Pillow==8.4.0
pkg_resources==0.0.0
pygame==2.1.0
pyparsing==3.0.6
python-dateutil==2.8.2
six==1.16.0
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
typing_extensions==4.1.1